pyproject.toml
[tool.poetry]
name = "task-4"
version = "0.1.0"
description = ""
authors = ["yaaa3ser <yasserissa1411@gmail.com>"]
readme = "README.md"
packages = [{include = "task_4"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
drf-yasg = {extras = ["validation"], version = "^1.21.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
